refactor: split adr-status-sync into a sibling module, track all first-party roots
Two more from the Codex review on #667.

**P1 — stop growing an already-exempted file.** The gate added 301 lines to
`scripts/check_ai_readiness.py`, taking it from 2,605 to 2,906 against a
documented 2,000-line hard cap that only passes because the file is in
`LARGE_FILE_ALLOWLIST` — which AGENTS.md is explicit is "not a way to
silently exempt a new file". Moved to `scripts/adr_status_sync.py`, a
pure-stdlib leaf module in the same shape as the other importable helpers
here (`evidence_tiers.py`, `platform_capabilities.py`). It owns the ADR
Status *parsing* helpers too, so `adr-index-nav-sync` reads them from there
instead of keeping a second copy, and the direction stays acyclic.
`check_ai_readiness.py` now ends at 2,563 lines — 42 fewer than before this
PR started, rather than 301 more.

**P2 — a Status names first-party files outside the package.** ADR-037's
already names `tests/test_cli_contract.py`, which the old regex forced under
`abicheck/` and dropped, so the receipt would never have warned on it. Any
`FIRST_PARTY_PY_ROOTS` tree now counts. The root list is a local literal to
keep the module a leaf, with a test asserting it agrees with
`check_ai_readiness.py`'s own — a new root can't go untracked silently.
Shorthand still expands against the package only, so a `tests/...` path
cannot anchor it into the wrong tree.
