+++ scripts/CLAUDE.md
@@ -11,6 +11,7 @@ Each must run with Python 3.10+ and the package installed in dev mode
 | `verify.py` | **The** verification orchestrator (CLAUDE.md "M0-3"). Runs a named `--profile {fast,pr,full}` bundle of the checks below through one step catalog, so `pixi run check`, `.pre-commit-config.yaml`, and `.github/workflows/ci.yml` can't independently drift from the real PR gate. `--only`/`--skip`/`--list`/`--json` for local iteration and machine-readable receipts. | Pixi (`pixi run check` = `--profile pr`), pre-commit (mypy + ai-readiness hooks), CI (`ai-readiness`/`lint-and-types`/`fair-metadata` jobs); manual. Mirrored in `tests/test_verify_profiles.py`. |
 | `gen_repo_facts.py` | Single source of truth for volatile repository facts (CLAUDE.md "M1-4") — project/latest-release version, example-catalog size, fast-lane test count, canonical Python. Writes `repo_facts.json`; `--check` fails on drift. `latest_release` is a maintained field (bump it when cutting a release), not auto-derived from git tags (unreliable on a shallow CI checkout). | `verify.py --only repo-facts` (CI `ai-readiness` job); manual. |
 | `check_ai_readiness.py` | AI-readiness gate (file size, CLAUDE.md coverage, test ratio, ChangeKind invariants, mypy baseline drift, import cycles, test-assertion density, ADR index/nav + status sync). | CI (`ai-readiness`) and `pre-commit`, both via `verify.py --only ai-readiness`. Exits 1 on errors. |
+| `adr_status_sync.py` | The `adr-status-sync` gate, split out of `check_ai_readiness.py` (already past the 2000-line hard cap): an ADR's own `**Status:**` line and its row in `docs/contribute/adr/index.md` may not *contradict* each other, an optional `**Verified:** <ref>@<sha> on <date>` receipt must parse, and commits after that sha touching a first-party file the Status names raise a WARN — the only check here that catches *document-vs-code* drift. Also owns the ADR Status parsing helpers (`_ADR_FILE_RE`, `_adr_status_text`) that `check_ai_readiness.py`'s own `adr-index-nav-sync` reuses. Pure-stdlib, importable. | imported (by `check_ai_readiness.py`); gated via `tests/test_ai_readiness.py`. |
 | `check_fp_rate.py` | False-positive/false-negative gate for public-surface scoping (ADR-024 §7). Labelled `(old, new)` corpus; baselines FP=0/FN=0. Cases are tagged by scoping *axis* (`CASE_CATEGORY`): `--json` carries a `by_category` breakdown and `--markdown` renders a per-axis table for CI step-summary / release-over-release trend reading. | CI (`ai-readiness`). Mirrored in `tests/test_fp_rate_gate.py`. |
 | `check_tier_accuracy.py` | Per-evidence-tier accuracy gate — *what each level buys*. One labelled logical change per case, projected down to what each tier observes (L0 symbols → L1 debug → L2 headers → L3 build) and run through `compare`; verdicts collapse to a 3-band ordinal (non-breaking/risk/breaking). Measures per-tier **over-call (FP)** vs **under-call (FN)** and which transition removes each. Gates: top-tier correctness + under-call monotonicity (more evidence never hides a break — ADR-028 D3). `--markdown`/`--json` emit the per-tier matrix. | CI (`ai-readiness`, + step-summary artifact). Mirrored in `tests/test_tier_accuracy_gate.py`. |
 | `measure_contract_shadow.py` | ADR-049 Phase 3's shadow contract-evaluator measurement **and** gate. Runs `check_fp_rate.py`'s labelled corpus **plus `contract_platform_corpus.py`'s platform/shape lanes** through `compare(..., contract_evaluation=True)` in all three contract domains and reports the phase's four "Measure:" quantities (delta matrix by old/new decision, unresolved rate by provider/domain/platform **and by Phase 6 lane**, proven public-break losses, proven false-positive reductions), naming the Gate lanes it deliberately cannot reach. Gates on three zero baselines: a proven public-break loss, a delta citing no (or a dangling) evidence record, and a finding with no recorded decision. `--markdown`/`--json`. | Manual / CI step summary. Mirrored in `tests/test_contract_shadow_measurement.py`. |
